eth, miner: move remote agent to it own package
